pyproject.toml
[tool.poetry]
name = "Mise en Place"
version = "0.1.0"
description = "/Users/kjaymiller/pinboard_daily/.direnv/python-3.10.5/bin/python /Users/kjaymiller/pinboard_daily/dashboard.py"
authors = ["Jay Miller <kjaymiller@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.1.18"
rich = "^12.5.1"
pinboard = "^2.1.8"
prompt-toolkit = "^3.0.30"
markdown2 = "^2.4.3"
feedparser = "^6.0.10"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
